Menos warnings

+++ Kernel/interruptions/intHandlers.c
@@ -175,7 +175,7 @@ int int80Handler(uint64_t rdi, uint64_t rsi, uint64_t rdx, uint64_t rcx) {
 			return getHeapSize(getCurrentPID());
 			break;
 		case 45: //getHeapBase
-			return getHeapBase(getCurrentPID());
+			return (uint64_t)getHeapBase(getCurrentPID());
 			break;
 		case 46: //killProcess
 			killProcess((int)rsi, -1);
